restore single plugin class
