+++ tests/SharedPluginsAndCodeactivites/AccountChainPostPlugin.cs
@@ -3,8 +3,7 @@ namespace DG.Some.Namespace {
     using System;
     using Microsoft.Xrm.Sdk;
     using DG.XrmFramework.BusinessDomain.ServiceContext;
-    using static SharedPluginsAndCodeactivites.Utility.Enums;
-
+    
     public class AccountChainPostPlugin : Plugin {
 
         public AccountChainPostPlugin()
