+++ tests/SharedPluginsAndCodeactivites/AccountPostPlugin.cs
@@ -3,8 +3,6 @@ namespace DG.Some.Namespace {
     using System;
     using Microsoft.Xrm.Sdk;
     using DG.XrmFramework.BusinessDomain.ServiceContext;
-    using static SharedPluginsAndCodeactivites.Utility.Enums;
-
     public class AccountPostPlugin : Plugin {
 
         public AccountPostPlugin()
