+++ tests/SharedPluginsAndCodeactivites/ContactPostPlugin.cs
@@ -2,8 +2,6 @@
     using System;
     using Microsoft.Xrm.Sdk;
     using DG.XrmFramework.BusinessDomain.ServiceContext;
-    using static SharedPluginsAndCodeactivites.Utility.Enums;
-
     public class ContactPostPlugin : Plugin 
     {
         public ContactPostPlugin()
