+++ tests/SharedPluginsAndCodeactivites/LeadRetrieveMultiplePlugin.cs
@@ -4,7 +4,6 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
-using static SharedPluginsAndCodeactivites.Utility.Enums;
 
 namespace DG.Some.Namespace
 {
