--- tests/SharedPluginsAndCodeactivites/Utility/AnyEntity.cs
@@ -1,14 +1,0 @@
-﻿namespace DG.Some.Namespace
-{
-    using Microsoft.Xrm.Sdk;
-
-    // StepConfig           : className, ExecutionStage, EventOperation, LogicalName
-    // ExtendedStepConfig   : Deployment, ExecutionMode, Name, ExecutionOrder, FilteredAttributes, UserContext
-    // ImageTuple           : Name, EntityAlias, ImageType, Attributes
-    #region PluginStepConfig made by Delegate A/S
-
-    class AnyEntity : Entity {
-        public AnyEntity() : base("") { }
-    }
-    #endregion
-}
